+++ gatewayservice/gateway-service.js
@@ -14,7 +14,7 @@ const llmServiceUrl = process.env.LLM_SERVICE_URL || 'http://localhost:8003';
 const authServiceUrl = process.env.AUTH_SERVICE_URL || 'http://localhost:8002';
 const userServiceUrl = process.env.USER_SERVICE_URL || 'http://localhost:8001';
 const gameServiceUrl = process.env.GAME_SERVICE_URL || 'http://localhost:8004';
-const wikidataServiceUrl = process.env.WIKIDATA_SERVICE_URL || 'http://localhost:8005';
+const wikidataServiceUrl = process.env.WIKIDATA_SERVICE_URL || 'http://localhost:3005';
 
 app.use(cors());
 app.use(express.json());
@@ -71,6 +71,16 @@ app.post('/addMatch', async (req, res) => {
 app.post('/endMatch', async (req, res) => {
   try {
     const userResponse = await axios.post(gameServiceUrl+'/endMatch', req.body);
+    res.json(userResponse.data);
+  } catch (error) {
+    res.status(error.response.status).json({ error: error.response.data.error });
+  }
+});
+
+
+app.post('/addQuestions', async (req, res) => {
+  try {
+    const userResponse = await axios.post(wikidataServiceUrl+'/addQuestions', req.body);
     res.json(userResponse.data);
   } catch (error) {
     res.status(error.response.status).json({ error: error.response.data.error });
